addresses review:
- Moves the methods from MultipleStoresTestRunner into MultipleStoresIT
  and removes MultipleStoresTestRunner
- Removes unneccessary start of mini cluster in MetaMultipleStoresIT
